blueprints_for_text_analysis/ch03_Scraping_Websites/requirements_03.txt
bs4
feedparser
isort
readability-lxml
scrapy
tqdm
watermark
xmltodict
